requirements.txt
audioread==3.0.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
contourpy==1.3.0
cycler==0.12.1
decorator==5.1.1
distlib==0.3.8
et-xmlfile==1.1.0
fastexcel==0.11.6
filelock==3.16.1
fonttools==4.54.1
fsspec==2024.10.0
huggingface-hub==0.26.2
idna==3.10
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.7
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.43.0
MarkupSafe==3.0.2
matplotlib==3.9.2
mpmath==1.3.0
msgpack==1.1.0
networkx==3.4.2
numba==0.60.0
numpy==1.26.4
openpyxl==3.1.5
packaging==24.1
pandas==2.2.2
pillow==11.0.0
pip==24.3.1
platformdirs==4.2.2
polars==1.0.0
pooch==1.8.2
pyarrow==17.0.0
pycparser==2.22
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
requests==2.32.3
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
six==1.16.0
soundfile==0.12.1
soxr==0.5.0.post1
sympy==1.13.3
threadpoolctl==3.5.0
timm==1.0.11
torch==2.2.2
torchvision==0.17.2
tqdm==4.67.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
virtualenv==20.26.3
XlsxWriter==3.2.0
